# core-lending-mortgages-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-mortgages
  description: Mortgages Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/mortgages
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/mortgage-applications/{mortgageApplicationId}/appraisals/{appraisalId}:
    get:
      tags:
        - MortgageAppraisal
      summary: Get an appraisal record by ID
      operationId: getById_10
      parameters:
        - name: mortgageApplicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: appraisalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgageAppraisalDTO'
    put:
      tags:
        - MortgageAppraisal
      summary: Update an appraisal record
      operationId: update_10
      parameters:
        - name: mortgageApplicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: appraisalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgageAppraisalDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgageAppraisalDTO'
    delete:
      tags:
        - MortgageAppraisal
      summary: Delete an appraisal record
      operationId: delete_10
      parameters:
        - name: mortgageApplicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: appraisalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/mortgage-applications/{mortgageApplicationId}/appraisals:
    get:
      tags:
        - MortgageAppraisal
      summary: List or search appraisals for a mortgage application
      operationId: findAll_10
      parameters:
        - name: mortgageApplicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.propertyId
          in: query
          description: Exact filter for property id
          required: false
          schema:
            type: string
        - name: filters.mortgageApplicationId
          in: query
          description: Exact filter for mortgage application id
          required: false
          schema:
            type: string
        - name: filters.appraiserName
          in: query
          description: Exact filter for appraiser name
          required: false
          schema:
            type: string
        - name: filters.licenseNumber
          in: query
          description: Exact filter for license number
          required: false
          schema:
            type: string
        - name: filters.appraisalType
          in: query
          description: Exact filter for appraisal type
          required: false
          schema:
            type: string
            enum:
              - FULL_APPRAISAL
              - DRIVE_BY
              - AUTOMATED
              - DESKTOP
              - OTHER
        - name: filters.marketValue
          in: query
          description: Exact filter for market value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[marketValue].from
          in: query
          description: Filter market value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[marketValue].to
          in: query
          description: Filter market value to value
          required: false
          schema:
            type: string
        - name: filters.rentalValue
          in: query
          description: Exact filter for rental value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[rentalValue].from
          in: query
          description: Filter rental value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[rentalValue].to
          in: query
          description: Filter rental value to value
          required: false
          schema:
            type: string
        - name: filters.replacementCost
          in: query
          description: Exact filter for replacement cost
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[replacementCost].from
          in: query
          description: Filter replacement cost from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[replacementCost].to
          in: query
          description: Filter replacement cost to value
          required: false
          schema:
            type: string
        - name: filters.landValue
          in: query
          description: Exact filter for land value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[landValue].from
          in: query
          description: Filter land value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[landValue].to
          in: query
          description: Filter land value to value
          required: false
          schema:
            type: string
        - name: filters.buildingValue
          in: query
          description: Exact filter for building value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[buildingValue].from
          in: query
          description: Filter building value from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[buildingValue].to
          in: query
          description: Filter building value to value
          required: false
          schema:
            type: string
        - name: filters.propertyCondition
          in: query
          description: Exact filter for property condition
          required: false
          schema:
            type: string
            enum:
              - EXCELLENT
              - GOOD
              - FAIR
              - POOR
        - name: filters.locationRating
          in: query
          description: Exact filter for location rating
          required: false
          schema:
            type: string
            enum:
              - A
              - B
              - C
              - D
              - E
        - name: filters.comparableProperties
          in: query
          description: Exact filter for comparable properties
          required: false
          schema:
            type: string
        - name: filters.appraisalDate
          in: query
          description: Exact filter for appraisal date
          required: false
          schema:
            type: string
        - name: filters.expiryDate
          in: query
          description: Exact filter for expiry date
          required: false
          schema:
            type: string
        - name: filters.requiresRepairs
          in: query
          description: Exact filter for requires repairs
          required: false
          schema:
            type: boolean
        - name: filters.requiredRepairs
          in: query
          description: Exact filter for required repairs
          required: false
          schema:
            type: string
        - name: filters.repairCost
          in: query
          description: Exact filter for repair cost
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[repairCost].from
          in: query
          description: Filter repair cost from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[repairCost].to
          in: query
          description: Filter repair cost to value
          required: false
          schema:
            type: string
        - name: filters.assumptions
          in: query
          description: Exact filter for assumptions
          required: false
          schema:
            type: string
        - name: filters.limitations
          in: query
          description: Exact filter for limitations
          required: false
          schema:
            type: string
        - name: filters.methodology
          in: query
          description: Exact filter for methodology
          required: false
          schema:
            type: string
        - name: filters.comments
          in: query
          description: Exact filter for comments
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseMortgageAppraisalDTO'
    post:
      tags:
        - MortgageAppraisal
      summary: Create a new appraisal record
      operationId: create_10
      parameters:
        - name: mortgageApplicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MortgageAppraisalDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MortgageAppraisalDTO'
components:
  schemas:
...
    MortgageAppraisalDTO:
      required:
        - appraisalDate
        - appraisalType
        - appraiserName
        - licenseNumber
        - locationRating
        - marketValue
        - mortgageApplicationId
        - propertyCondition
        - propertyId
        - requiresRepairs
      type: object
      properties:
        appraisalId:
          type: string
          format: uuid
          readOnly: true
        propertyId:
          type: string
          format: uuid
        mortgageApplicationId:
          type: string
          format: uuid
        appraiserName:
          maxLength: 100
          minLength: 0
          type: string
        licenseNumber:
          maxLength: 50
          minLength: 0
          type: string
        appraisalType:
          type: string
          enum:
            - FULL_APPRAISAL
            - DRIVE_BY
            - AUTOMATED
            - DESKTOP
            - OTHER
        marketValue:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        rentalValue:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        replacementCost:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        landValue:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        buildingValue:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        propertyCondition:
          type: string
          enum:
            - EXCELLENT
            - GOOD
            - FAIR
            - POOR
        locationRating:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - E
        comparableProperties:
          maxLength: 5000
          minLength: 0
          type: string
        appraisalDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        requiresRepairs:
          type: boolean
        requiredRepairs:
          maxLength: 2000
          minLength: 0
          type: string
        repairCost:
          maximum: 999999999.99
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        assumptions:
          maxLength: 2000
          minLength: 0
          type: string
        limitations:
          maxLength: 2000
          minLength: 0
          type: string
        methodology:
          maxLength: 1000
          minLength: 0
          type: string
        comments:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseMortgageAppraisalDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/MortgageAppraisalDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.